IPv6 support, use standard classes for universal IPParser

+++ udf/src/main/java/com/dmittov/mrdemo/udf/GeobaseLoader.java
@@ -1,11 +1,12 @@
 package com.dmittov.mrdemo.udf;
 
 import java.io.IOException;
+import java.math.BigInteger;
 import java.util.Map;
 
 /**
  * Created by mittov on 13/01/2017.
  */
 public interface GeobaseLoader {
-    Map<Long, String> loadGeoBase() throws IOException;
+    Map<BigInteger, String> loadGeoBase() throws IOException;
 }
